Basic selenium test on instrumented client code with coverage report
All client JS was instrumented with istanbul and replaced. The original
files are in src/js/client/orig.

To instrument code with istanbul, run the instrument command:
/node_modules/.bin/istanbul instrument --embed-sources <source_dir> -o <output_dir>

The coverage reporting relies on a server using istanbul-middleware.
To use, checkout and bulid this project https://github.com/aaewong/istanbul-report-server

The selenium test grabs the window.__coverage__ object that contains the
coverage data from the execution and does the following:
- saves it to a local file at coverage/coverage.json
- sends it in an HTTP POST to istanbul-report-server at localhost:8888

Running the tests:
1. Start the istanbul-report-server
2. Start the application dev server: npm run server
3. Run the tests: npm test
4. Go to localhost:8888/coverage to view the report